V0.19

+++ root/includes/acp/acp_users_membership.php
@@ -1,21 +1,27 @@
 <?php
 
 /**
- * @author Action Replay
- * @copyright 2011
- */
+* @package acp_users_membership.php
+* @copyright (c) DougA http://action-replay.co.uk 2011
+* @license http://opensource.org/licenses/gpl-license.php GNU Public License
+*
+*/
+if (!defined('IN_PHPBB'))
+{
+	exit;
+}
 
 	include($phpbb_root_path . 'includes/functions_user.' . $phpEx);
 	include($phpbb_root_path . 'includes/functions_membership.' . $phpEx);
 
-	$user->add_lang(array('mods/application'));    
+	$user->add_lang(array('mods/application'));	
 
-    $delete_members=request_var('delete','');
-    if ($delete_members)
-    {
+	$delete_members=request_var('delete','');
+	if ($delete_members)
+	{
 		if (confirm_box(true))
 		{
-            group_user_del($config['ms_subscription_group'], $user_id);
+			group_user_del($config['ms_subscription_group'], $user_id);
 			$sql = 'DELETE FROM ' . MEMBERSHIP_TABLE . ' WHERE group_id=' . $config['ms_subscription_group'] . ' AND user_id =' . $user_id;
 			$result = $db->sql_query($sql);
 			trigger_error($user->lang['USER_DELETED'] . adm_back_link($this->u_action));
@@ -29,48 +35,48 @@
 				'action'		=> $action,
 				'update'		=> true,
 				'delete'		=> 1,
-                ))
+				))
 			);
 		}
-    }                
-    $sql_array = array(
-        'SELECT'    => 'm.*, u.username_clean',
-        'FROM'      => array(
-            MEMBERSHIP_TABLE=> 'm',
-            ),
-        'LEFT_JOIN' => array(
-            array(
-                'FROM'  => array(USERS_TABLE => 'u'),
-                'ON'    => 'u.user_id = m.associate_id'
-            ),
-        ),
-        'WHERE'     	=>  'm.user_id = '. $user_id . ' AND m.group_id = '. $config['ms_subscription_group'],
-    );
-    $sql=$db->sql_build_query('SELECT', $sql_array);
-    $result = $db->sql_query($sql);
-    $row = $db->sql_fetchrow($result);
-    
-    $db->sql_freeresult($result);
+	}				
+	$sql_array = array(
+		'SELECT'	=> 'm.*, u.username_clean',
+		'FROM'	=> array(
+			MEMBERSHIP_TABLE=> 'm',
+			),
+		'LEFT_JOIN' => array(
+			array(
+				'FROM'  => array(USERS_TABLE => 'u'),
+				'ON'	=> 'u.user_id = m.associate_id'
+			),
+		),
+		'WHERE'		=>  'm.user_id = '. $user_id . ' AND m.group_id = '. $config['ms_subscription_group'],
+	);
+	$sql=$db->sql_build_query('SELECT', $sql_array);
+	$result = $db->sql_query($sql);
+	$row = $db->sql_fetchrow($result);
+	
+	$db->sql_freeresult($result);
 
 	$not_member = empty($row);
 
-    if ($not_member)
+	if ($not_member)
 	{
 		$row['membership_no'] = '';
 		$row['username_clean'] = '';
 		$row['renewal_date']=strtotime('+3 month');
 		$row['datepaid']=time();
-    }
+	}
 	$data = array(
-		'membership_no'       => request_var('membership_no', $row['membership_no']),
-		'associate_name'        => request_var('associate_name', $row['username_clean'].''),
+		'membership_no'		=> request_var('membership_no', $row['membership_no']),
+		'associate_name'	=> request_var('associate_name', $row['username_clean'].''),
 	);
-    $data['rday_day']           = request_var('rday_day', date('j',$row['renewal_date']));
-    $data['rday_month']         = request_var('rday_month', date('n',$row['renewal_date']));
-    $data['rday_year']          = request_var('rday_year', date('Y',$row['renewal_date']));
-    $data['pday_day']           = request_var('pday_day', date('j',$row['datepaid']));
-    $data['pday_month']         = request_var('pday_month', date('n',$row['datepaid']));
-    $data['pday_year']          = request_var('pday_year', date('Y',$row['datepaid']));
+	$data['rday_day']		= request_var('rday_day', date('j',$row['renewal_date']));
+	$data['rday_month']		= request_var('rday_month', date('n',$row['renewal_date']));
+	$data['rday_year']		= request_var('rday_year', date('Y',$row['renewal_date']));
+	$data['pday_day']		= request_var('pday_day', date('j',$row['datepaid']));
+	$data['pday_month']		= request_var('pday_month', date('n',$row['datepaid']));
+	$data['pday_year']		= request_var('pday_year', date('Y',$row['datepaid']));
 
 	if ($submit)
 	{
@@ -84,22 +90,22 @@
 			'pday_year'		=> array('num', true, 1901, gmdate('Y', time())+10),
 		));
 
-        if ($data['membership_no'] != $row['membership_no'])
-        {
+		if ($data['membership_no'] != $row['membership_no'])
+		{
 			if (!empty($data['membership_no']))
 			{
 				$sql_array = array(
-					'SELECT'    => 'u.username_clean',
-					'FROM'      => array(
+					'SELECT'	=> 'u.username_clean',
+					'FROM'	=> array(
 						MEMBERSHIP_TABLE=> 'm',
 						),
 					'LEFT_JOIN' => array(
 						array(
 							'FROM'  => array(USERS_TABLE => 'u'),
-							'ON'    => 'u.user_id = m.user_id'
+							'ON'	=> 'u.user_id = m.user_id'
 							),
 						),
-					'WHERE'     	=>  "membership_no={$data['membership_no']}",
+					'WHERE'		=>  "membership_no={$data['membership_no']}",
 					);
 				$sql=$db->sql_build_query('SELECT', $sql_array);
 				$result =$db->sql_query($sql);
@@ -113,33 +119,33 @@
 					$update_ary['membership_no']=$data['membership_no'];
 				}
 			}
-        }
-        
-        if ($data['associate_name'] != $row['username_clean'])
-        {
-            $assoc_error = validate_associate($data['associate_name'], $associate_id);
-            if (!empty($assoc_error))
-            {
-                $error[] = $assoc_error;
-            }
-        }
-        
-        if (checkdate($data['rday_month'],$data['rday_day'],$data['rday_year']))
-        {
-            $update_ary['renewal_date'] =    mktime(0,0,0,$data['rday_month'],$data['rday_day'],$data['rday_year']);
-        }
-        else
-        {
-            $error[] = 'INVALID_DATE';            
-        }
-        if (checkdate($data['pday_month'],$data['pday_day'],$data['pday_year']))
-        {
-            $update_ary['datepaid'] =    mktime(0,0,0,$data['pday_month'],$data['pday_day'],$data['pday_year']);
-        }
-        else
-        {
-            $error[] = 'INVALID_DATE';            
-        }
+		}
+		
+		if ($data['associate_name'] != $row['username_clean'])
+		{
+			$assoc_error = validate_associate($data['associate_name'], $associate_id);
+			if (!empty($assoc_error))
+			{
+				$error[] = $assoc_error;
+			}
+		}
+		
+		if (checkdate($data['rday_month'],$data['rday_day'],$data['rday_year']))
+		{
+			$update_ary['renewal_date'] =	mktime(0,0,0,$data['rday_month'],$data['rday_day'],$data['rday_year']);
+		}
+		else
+		{
+			$error[] = 'INVALID_DATE';			
+		}
+		if (checkdate($data['pday_month'],$data['pday_day'],$data['pday_year']))
+		{
+			$update_ary['datepaid'] =	mktime(0,0,0,$data['pday_month'],$data['pday_day'],$data['pday_year']);
+		}
+		else
+		{
+			$error[] = 'INVALID_DATE';			
+		}
 
 		if (!check_form_key($form_name))
 		{
@@ -150,10 +156,10 @@
 		{
 			update_membership($config['ms_subscription_group'], $user_id, $update_ary);
 
-            if ($data['associate_name'] != $row['username_clean'])
-            {
-                process_associate($user_id, $associate_id);                
-            }
+			if ($data['associate_name'] != $row['username_clean'])
+			{
+				process_associate($user_id, $associate_id);				
+			}
 			trigger_error($user->lang['USER_PROFILE_UPDATED'] . adm_back_link($this->u_action . '&amp;u=' . $user_id));
 		}
 		// Replace "error" strings with their real, localised form
@@ -196,7 +202,7 @@
 	unset($now);
 
 	$template->assign_vars(array(
-        'NOT_MEMBER'            => $not_member,
+		'NOT_MEMBER'			=> $not_member,
 		'MEMBERSHIP_NO'			=> $data['membership_no'],
 		'ASSOCIATE_NAME'			=> $data['associate_name'],
 		'S_RENEWAL_DAY_OPTIONS'	=> $s_renewal_day_options,
